+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Added a list of failed contacts to sendouts that have failed ([#311](https://github.com/putyourlightson/craft-campaign/issues/311)).
 
+### Changed
+- Replaced all instances of `AdminTable` with `VueAdminTable`.
+
 ## 2.0.0-beta.2 - 2022-04-17
 ### Added
 - Added an "Export to CSV" button to all datatables in reports ([#245](https://github.com/putyourlightson/craft-campaign/issues/245)).